Move to strategy-style IntegratedTests, to be run through JUnit
